requirements.txt
ultralytics
torch
opencv-python
numpy
PyYAML
tqdm
matplotlib
